trading_crypto/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cycler==0.10.0
gevent==20.6.2
greenlet==0.4.16
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kiwisolver==1.2.0
matplotlib==3.2.2
multidict==4.7.6
numpy==1.19.0
pika==1.1.0
psutil==5.7.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
redis==3.5.3
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.17
txaio==20.4.1
websocket-client==0.57.0
yarl==1.4.2
zope.event==4.4
zope.interface==5.1.0
